updating packages and trying to fix build
