+++ package.json
@@ -15,34 +15,33 @@
   },
   "dependencies": {
     "@auth/dynamodb-adapter": "^2.0.0",
-    "@aws-sdk/client-dynamodb": "^3.556.0",
-    "@aws-sdk/lib-dynamodb": "^3.556.0",
-    "@next-auth/dynamodb-adapter": "^3.0.2",
-    "asdf": "^1.2.2",
-    "axios": "^1.6.0",
-    "govuk-frontend": "^4.7.0",
+    "@aws-sdk/client-dynamodb": "^3.569.0",
+    "@aws-sdk/lib-dynamodb": "^3.569.0",
+    "asdf": "^1.2.3",
+    "axios": "^1.6.8",
+    "govuk-frontend": "^5.3.1",
     "jsonwebtoken": "^9.0.2",
     "jwks-rsa": "^3.1.0",
-    "next": "^13.5.6",
-    "next-auth": "^4.24.5",
-    "nhsuk-frontend": "^7.0.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "next": "^14.2.3",
+    "next-auth": "^5.0.0-beta.17",
+    "nhsuk-frontend": "^8.1.1",
+    "react": "18.3.1",
+    "react-dom": "18.3.1"
   },
   "cacheDirectories": [
     "node_modules/",
     ".next/cache/"
   ],
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.0.1",
-    "@testing-library/react": "^14.0.0",
+    "@testing-library/jest-dom": "^6.4.5",
+    "@testing-library/react": "^15.0.6",
     "@types/jsonwebtoken": "^9.0.6",
     "gulp": "^5.0.0",
     "gulp-sass": "^5.1.0",
-    "jest": "^29.6.3",
-    "jest-environment-jsdom": "^29.6.3",
-    "react-router-dom": "^6.15.0",
-    "sass": "^1.64.2",
-    "typescript": "5.3.2"
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "react-router-dom": "^6.23.0",
+    "sass": "^1.77.0",
+    "typescript": "5.4.5"
   }
 }
