+++ src/app/onwardreferral/page.jsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React from "react";
-import "src/app/styles/css/sass.css";
+import "../styles/css/sass.css";
 import { useSession } from "next-auth/react";
 import { useInactivity } from "@/app/context/AutoSignOutProvider";
 import Header from "@/app/components/Header";
